requirements.txt
streamlit
openai
mcp
python-dotenv
beautifulsoup4
chromadb
sentence-transformers
pandas
matplotlib
seaborn
python-pptx
python-docx
PyMuPDF
google-auth-oauthlib
google-api-python-client
requests
httpx